[DEV-5427] Move some files around and only return non-zero outlay and obligation
