+++ usaspending_api/disaster/tests/fixtures/disaster_account_data.py
@@ -1,4 +1,3 @@
-import json
 import pytest
 
 from datetime import datetime
@@ -269,22 +268,3 @@ def disaster_account_data():
         obligations_incurred_by_program_object_class_cpe=100,
         gross_outlay_amount_by_program_object_class_cpe=100000,
     )
-
-
-class Helpers:
-    @staticmethod
-    def post_for_count_endpoint(client, url, def_codes=None):
-        if def_codes:
-            request_body = json.dumps({"filter": {"def_codes": def_codes}})
-        else:
-            request_body = json.dumps({"filter": {}})
-        resp = client.post(url, content_type="application/json", data=request_body)
-        return resp
-
-
-@pytest.fixture
-def helpers():
-    return Helpers
-
-
-__all__ = ["disaster_account_data", "helpers"]
